Rolling sent to server, updating socket data on the server, rolling page is in a "full testable state"
We really need to have a way to get player data (and by player data i
mean game data for a given player) from the server. This will probably
need to be stored in the database in the games collection or something
and we can run some query on it. I also need to deal with property data
and the games.