Trying to fix app.js database queries but failing
The database queries need to always have callbacks. If you don't have a
callback, the return statement will fire and you are almost guarnateed
to have a null value. Each of the query functions now takes in a
callback funtion, but I'm still umable to get gameQuery to find
anything. It could be the way I am joining games or that we aren't
updating stuff properly. I really can't figure it out right now :(